requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.0
jinja2==3.1.4
psycopg[binary]==3.2.3
pydantic==2.10.6
python-dotenv==1.0.1
paho-mqtt==1.6.1
python-multipart==0.0.20
